eventInfo Css, Landing Logo, Delete Alert

+++ client/App.jsx
@@ -370,7 +370,6 @@ class App extends React.Component {
       params,
     })
       .then((res) => {
-        // console.log(res.data);
         this.setState({
           eventInfoAccess: res.data.access,
           eventInfo: res.data.eventInfo[0],
@@ -391,9 +390,7 @@ class App extends React.Component {
   }
 
   handleRemoveGuest(event) {
-    console.log("REMOVING")
-    // console.log(displayName);
-    // console.log(this.state.eventId);
+    // console.log("REMOVING")
     const eventId = event.target.getAttribute('data-id')
     axios.delete('/pending', {
       params: {
@@ -402,7 +399,6 @@ class App extends React.Component {
       },
     })
       .then(() => {
-        // console.log(res.data); //<-------------------------Remove
         this.openEventInfoModal(eventId);
       })
       .catch((err) => {
@@ -418,7 +414,6 @@ class App extends React.Component {
       eventId: this.state.eventId,
     })
       .then((res) => {
-        // console.log(res.data); //<-------------------------Remove
         this.openEventInfoModal(this.state.eventId);
       })
       .catch((err) => {
@@ -436,7 +431,6 @@ class App extends React.Component {
       eventId,
     })
       .then((res) => {
-        // console.log(res.data); //<-------------------------Remove
         this.openEventInfoModal(eventId);
       })
       .catch((err) => {
@@ -480,18 +474,21 @@ class App extends React.Component {
   }
 
   handleDeleteHostEvent(uniqueId) {
-    const params = { uniqueId };
-    axios.delete('/event', {
-      params,
-    })
-      .then((response) => {
-        this.getAllEvents();
-        this.closeEventInfoModal();
-        this.getEventsForDashboard();
+    let retVal = confirm('Are you sure you want to delete this event?');
+    if (retVal === true) {
+      const params = { uniqueId };
+      axios.delete('/event', {
+        params,
       })
-      .catch((err) => {
-        console.log('didn\'t work from axios', err);
-      });
+        .then((response) => {
+          this.getAllEvents();
+          this.closeEventInfoModal();
+          this.getEventsForDashboard();
+        })
+        .catch((err) => {
+          console.log('didn\'t work from axios', err);
+        });
+    } 
   }
 
   handleGuestSubmit(event) {
